refactor: découpage Partnership + corrections mineures audit
- PartnershipCarouselSection: 1058 → 237 lignes (-78%)
  → 5 sous-composants dans carousel/ (types, data, slides, controls, header)
- PartnershipVisionSection: 943 → 132 lignes (-86%)
  → 6 sous-composants (Header, Timeline, Production, Verticales, Setup, Editorial)
  + visionData.ts pour les types et données partagées
- package.json: nom "origines-media", version "1.0.0"
- index.css: nettoyage variables dark mode inutilisées

Build OK, 23 tests passent.
