+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "origines-media-front",
+  "name": "origines-media",
   "private": true,
   "version": "1.0.0",
   "type": "module",
