+++ src/components/partnership/carousel/index.ts
@@ -1,0 +1,5 @@
+export { default as CarouselSlideContent } from './CarouselSlideContent';
+export { default as CarouselHeader } from './CarouselHeader';
+export { default as CarouselControls, CarouselMobileNavigation, CarouselSlideIndicators, CarouselBottomControls } from './CarouselControls';
+export { slides } from './carouselSlidesData';
+export type { Slide } from './carouselTypes';
